pyproject.toml
[tool.poetry]
name = "hc_pynomad"
version = "0.1.0"
description = "Unofficial Python client for HashiCorp Nomad."
authors = ["Arden Shackelford <arden@ardens.tech>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9,<=4.0"
pydantic = "^1.9.0"
httpx = "^0.22.0"

[tool.poetry.group.testing.dependencies]
pytest-cov = "^4.0.0"
mypy = "^1.3.0"
respx = "^0.20.1"
ruff = "^0.0.265"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = ["__init__.py"]
